Add a bunch of name-and-date-only signs.
With this, every machine in the list I was given has a sign of some sort.
I still plan on doing write-ups for one or two more systems, but this is
a completion of the minimum.  (There are two HP85 signs because we don't
know whether the machine is an 85A or an 85B.)